fix(event and ruleset) : update DTOs and Dal

+++ src/main/java/com/filsanguinaire/tournament/dto/rules/AllowedInducementDTO.java
@@ -15,7 +15,7 @@ public class AllowedInducementDTO {
     
     private String name;
     
-    private Integer minQty;
+    private short minQty;
     
-    private Integer maxQty;
+    private short maxQty;
 }
